Move InventoryGUIItemButton into src/, include/ files
